requirements.txt
comfyui>=0.9.0
numpy>=1.24
Pillow>=9.0
